refactor: 리뷰 사항 반영 (useEditPost_getPost 삭제, 로직 컴포넌트에 병합)
